requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
bcrypt
python-dotenv
requests